ensuring float the answer

+++ math/0x03-probability/exponential.py
@@ -16,7 +16,7 @@ class Exponential:
                 raise TypeError("data must be a list")
             if len(data) < 2:
                 raise ValueError("data must contain multiple values")
-            self.lambtha = sum(data)/len(data)
+            self.lambtha = float(sum(data)/len(data))
 
     def pmf(self, k):
         """Method pmf fr Poisson dist"""
